refactor: estrai domain/common.py da cli/common.py
